backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pvlib>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
redis>=5.0.0
httpx>=0.24.0
python-multipart
python-dotenv
geoalchemy2>=0.14.0
shapely>=2.0.0
